cdl: Track each VkQueue separately
Change class SubmitTracker into class Queue. Instead of 1 per device,
create a separate tracker for every VkQueue. This lets us separate
and order submissions so that it is easier to debug multi-queue hangs.
